MAGNOLIA-3614: executes a rudimentary jcr query, instead of the fake search we had so far.
git-svn-id: file:///home/migrator/subversion-migration/svn-repo-copy/community/magnolia/trunk@44274 a15fbcf9-bee4-0310-bafe-d02d26ad3b5f

+++ magnolia-ui-admincentral/src/main/java/info/magnolia/ui/admincentral/search/view/SearchResult.java
@@ -40,9 +40,9 @@ package info.magnolia.ui.admincentral.search.view;
  */
 public class SearchResult {
     private String queryTerm;
-    private int itemsFound;
+    private long itemsFound;
 
-    public SearchResult(String queryTerm, int itemsFound) {
+    public SearchResult(String queryTerm, long itemsFound) {
         this.queryTerm = queryTerm;
         this.itemsFound = itemsFound;
     }
@@ -51,7 +51,7 @@ public class SearchResult {
         return queryTerm;
     }
 
-    public int getItemsFound() {
+    public long getItemsFound() {
         return itemsFound;
     }
